ftp://ftp.krokus.ru/pub/OpenBSD/distfiles/xscreensaver-4.06.tar.gz -rw-r--r-- 1 zblaxell zblaxell 3209164 Dec 16  2002 xscreensaver-4.06.tar.gz ba2c14d632ee5d876e1d2b580e6aa6d67b813119  xscreensaver-4.06.tar.gz

+++ driver/remote.c
@@ -278,13 +278,13 @@ send_xscreensaver_command (Display *dpy, Atom command, long arg,
                     }
 
                 if (any && nhacks == 1)
-                  fprintf (stdout, " (hack #%d)\n", data[2]);
+                  fprintf (stdout, " (hack #%d)\n", (int) data[2]);
                 else if (any)
                   {
                     fprintf (stdout, " (hacks: ");
                     for (i = 0; i < nhacks; i++)
                       {
-                        fprintf (stdout, "#%d", data[2 + i]);
+                        fprintf (stdout, "#%d", (int) data[2 + i]);
                         if (i != nhacks-1)
                           fputs (", ", stdout);
                       }
